+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All notable changes to the Zowe Client Python SDK will be documented in this fil
 ## Recent Changes
 
 - BugFix: Validation of zowe.config.json file matching the schema [#192](https://github.com/zowe/zowe-client-python-sdk/issues/192)
+- Feature: Added method to load profile properties from environment variables